--- a/core/core.mk
+++ b/core/core.mk
@@ -15,7 +15,6 @@ PLATFORM_$(PLATFORM) := y
 PLATFORM_FLAVOR_$(PLATFORM_FLAVOR) := y
 
 $(call cfg-depends-all,CFG_PAGED_USER_TA,CFG_WITH_PAGER CFG_WITH_USER_TA)
-CFG_LTC_CRYPTO_GCM = $(CFG_CRYPTO_GCM)
 
 # Setup compiler for this sub module
 COMPILER_$(sm)		?= $(COMPILER)