… core/arch/arm/plat-hikey/conf.mk
… core/crypto/aes-gcm.c
… core/crypto/sub.mk
… core/include/crypto/aes-gcm.h
… core/include/crypto/aes-gcm_ext.h
… core/include/io.h
… core/lib/libtomcrypt/include/tomcrypt_custom.h
… core/lib/libtomcrypt/src/encauth/gcm/gcm_mult_h_arm_ce.c
… core/lib/libtomcrypt/src/encauth/gcm/sub.mk
… core/lib/libtomcrypt/src/encauth/sub.mk
… core/lib/libtomcrypt/src/tee_ltc_provider.c