Hardware optimized AES-GCM implementation. Here's some figures from testing on Hikey:

"default" means what's on master today.

command: `xtest --aes-perf -m ECB`
default:
min=99.166us max=153.334us mean=107.972us stddev=3.59459us (cv 3.32917%) (9.04455MiB/s)

command: `xtest --sha-perf -a SHA256`
default:
min=105.833us max=149.167us mean=113.857us stddev=3.38391us (cv 2.97209%) (8.57713MiB/s)

command: `xtest --aes-perf -m GCM`

default:
min=1445.83us max=1503.33us mean=1471.38us stddev=9.70065us (cv 0.659288%) (0.663704MiB/s)

LTC + GHASH (commit "LTC: add GHASH acceleration")
min=178.334us max=222.5us mean=188.389us stddev=4.36719us (cv 2.31818%) (5.18376MiB/s)

LTC + GHASH + PMULL (commit "hikey: CFG_HWSUPP_PMULL=y")
min=170.833us max=210.834us mean=172.781us stddev=3.58542us (cv 2.07512%) (5.65202MiB/s)

With this pull request:
min=98.333us max=127.5us mean=102.794us stddev=3.55137us (cv 3.45483%) (9.50017MiB/s)

There's obviously rooms for improvements for AES-ECB mode as AES-GCM is a bit faster. I suspect that our SHA-256 implementation can be improved too.