@jenswi-linaro 2017-10-31
Number for new AES-GCM implementation withouth the fused fused AES+PMULL

min=153.333us max=178.333us mean=155.11us stddev=3.2519us (cv 2.09652%) (6.29596MiB/s)