@jenswi-linaro 2017-10-31
Checkpatch fixes