@jenswi-linaro 2017-11-01
There's even a small gain in "core: arm: crypto: fix AES-GCM counter increase"
now AES-GCM on Hikey is:
min=100us max=124.167us mean=100.995us stddev=2.56812us (cv 2.54281%) (9.66937MiB/s)