@jforissier 2017-11-02
`Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey960)`

I noticed a strange thing: this PR seems to make ECB and SHA256 significantly slower on the a73 cores, but not on the a53 cores. See below. Any idea what might be happening?

```
# HiKey960 a73 cores only:
for i in 4 5 6 7; do echo 1 >/sys/devices/system/cpu/cpu$i/online; done
for i in 0 1 2 3; do echo 0 >/sys/devices/system/cpu/cpu$i/online; done
# HiKey960 a53 cores only:
for i in 0 1 2 3; do echo 1 >/sys/devices/system/cpu/cpu$i/online; done
for i in 4 5 6 7; do echo 0 >/sys/devices/system/cpu/cpu$i/online; done


### master, a73 cores:

xtest --aes-perf -m ECB
min=24.479us max=36.979us mean=24.8683us stddev=0.505735us (cv 2.03365%) (39.2693MiB/s)

xtest --sha-perf -a SHA256
min=26.562us max=38.542us mean=26.888us stddev=0.507852us (cv 1.88877%) (36.3196MiB/s)

xtest --aes-perf -m GCM
min=408.854us max=425.52us mean=413.601us stddev=1.9973us (cv 0.482905%) (2.36112MiB/s)


### master, a53 cores:

xtest --aes-perf -m ECB
min=59.895us max=94.271us mean=60.5282us stddev=1.28037us (cv 2.11532%) (16.134MiB/s)

xtest --sha-perf -a SHA256
min=64.583us max=100us mean=65.1378us stddev=1.4101us (cv 2.1648%) (14.9923MiB/s)

xtest --aes-perf -m GCM
min=934.375us max=982.812us mean=948.535us stddev=5.34257us (cv 0.563245%) (1.02955MiB/s)


### PR#1908 + CFG_HWSUPP_PMULL=y, a73 cores:

xtest --aes-perf -m ECB
min=32.291us max=45.834us mean=32.8579us stddev=0.521768us (cv 1.58795%) (29.7208MiB/s)

xtest --sha-perf -a SHA256
min=34.375us max=47.396us mean=34.9583us stddev=0.527139us (cv 1.50791%) (27.935MiB/s)

xtest --aes-perf -m GCM
min=32.291us max=48.437us mean=32.9785us stddev=0.540848us (cv 1.64%) (29.6121MiB/s)


### PR#1908 + CFG_HWSUPP_PMULL=y, a53 cores:

xtest --aes-perf -m ECB
min=59.895us max=103.646us mean=64.5185us stddev=1.96989us (cv 3.05322%) (15.1362MiB/s)

xtest --sha-perf -a SHA256
min=64.583us max=95.833us mean=68.9705us stddev=1.63784us (cv 2.3747%) (14.1591MiB/s)

xtest --aes-perf -m GCM
min=59.895us max=101.562us mean=64.2076us stddev=1.88134us (cv 2.93009%) (15.2095MiB/s)
```