@jbech-linaro 2017-11-06
FYI: OP-TEE Hikey auto builder - shows false positive, "OK" here but the logs shows errors (and the log from the script also says "failed", but still it will send an "OK" to GitHub. I thought I had this fixed, but apparently not).