@jenswi-linaro 2017-11-09
For "hikey: CFG_HWSUPP_PMULL=y":
actually this seems to be unintentionally applied to 960 already. I'll update the commit message to make that clear.

We'd better ask @ardbiesheuvel about "core: arm64: add mov_imm assembly macro". @ardbiesheuvel are you OK with a sign-off or would you rather ack it or something?

I'll add a function based interface for the crypto provider in this PR.