@jforissier 2017-11-15
@jenswi-linaro about Ard's S-o-b: I'm citing a maintainer on the LKML [1]:

"The last Signed-off-by should match the person posting the patch. It's a chain of custody for GPL-signoff purposes, not a 'partially-written-by'. If you want to credit [someone] you can add CC and they can Ack/Review your patch."

I think the same reasoning should apply here.

[1] https://lkml.org/lkml/2017/11/14/390