@ardbiesheuvel 2017-11-15
On 15 November 2017 at 07:56, Jérôme Forissier <notifications@github.com>
wrote:

> @jenswi-linaro <https://github.com/jenswi-linaro> about Ard's S-o-b: I'm
> citing a maintainer on the LKML [1]:
>
> "The last Signed-off-by should match the person posting the patch. It's a
> chain of custody for GPL-signoff purposes, not a 'partially-written-by'. If
> you want to credit [someone] you can add CC and they can Ack/Review your
> patch."
>
> I think the same reasoning should apply here.
>

​Works for me

Acked-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
​

> [1] https://lkml.org/lkml/2017/11/14/390
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/OP-TEE/optee_os/pull/1908#issuecomment-344513630>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AIDoWEhP9jkT90totG4c-_tTtCLUahYqks5s2pkjgaJpZM4QMkTK>
> .
>