@jenswi-linaro 2017-11-16
Rebased on master