@jforissier 2017-11-17
I am still concerned by "core: ltc: remove redundant aes-gcm implementation". It means it is not possible to use LibTomCrypt anymore for AES GCM, and as I said already it might be a problem. Some people may be using their own modified version of LTC, or may have audited LTC and do not want to use anything else...

More generally, and this relates to https://github.com/OP-TEE/optee_os/pull/1931 actually, the crypto API in `crypto.h` is not clean enough IMO:
- `core/arch/crypto/crypto.c` depends on `_CFG_CRYPTO` flags defined in LTC.
- The `crypto_authenc_*()` functions, although supposedly part of the API, cannot be overridden.

So how to fix this? Well it shouldn't be too hard hopefully:
1. Move the `_CFG_CRYPTO` flags out of LTC, from `core/lib/libtomcrypt/sub.mk` to `core/crypto/sub.mk`. They do not depend on a particular implementation.
2. Introduce `CFG_CRYPTO_GCM_CE` to select this implementation of `crypto_authenc_*()`, otherwise use LTC.

What do you think?