@jenswi-linaro 2017-11-17
OK, I see your point.

This means that we'll drop b1538e99efe1 ("core: crypto: merge aes-gcm.h and aes-gcm_ext.h") too.

The flags should probably go into `core/core.mk` instead since `sub.mk` shouldn't set flags because only the `sub.mk` parsed after that will see the changes.

How about instead of `CFG_CRYPTO_GCM_CE` have a `CFG_CRYPTO_AES_GCM_FROM_CRYPTOLIB` which would disable the internal AES-GCM implementation and expect the crypto library (LTC for now) to provide an AES-GCM implementation instead.

However, this doesn't sort it out completely, we have a third(!) AES-GCM implementation used by the pager too. I was planning to replace that implementation with the new AES-GCM implementation. The LTC implementation can't be used for this so we'll need a way to let the two implementations (LTC and new) coexist. One way would be to have a separate API for the new AES-GCM implementation with a good prefix instead of `crypto_`, and then just some wrappers to provide the required `crypto_aes_gcm_*()` functions when configured to do so. Do you have suggestion for the "good prefix"? I've run out of ideas for the moment.